Documentation Character

+++ src/Entity/Timers.cpp
@@ -18,7 +18,7 @@ void Timers::deleteExpired() {
 }
 
 bool Timers::isClocked(uint_least64_t entityId) {
-	for(unsigned int i = 0; i < mClocks.size(); ++i) {
+	for(size_t i = 0; i < mClocks.size(); ++i) {
 		if(mClocks[i].first == entityId) {
 			if(mClocks[i].second.getElapsedTime() > mMaxTime) {
 				mClocks.erase(mClocks.begin() + i);
